+++ src/include/Autorit.hpp
@@ -4,10 +4,10 @@
 
 class Autorit : public Edge{  
     private:
-        static int multiplyCost(int* cost);
+        static float multiplyCost(float* cost);
 
     public:
-        Autorit(Node* from, Node* to, int cost);
+        Autorit(Node* from, Node* to, float cost);
         std::string getType() override;
         bool operator==(Edge* rhs) const override;
 };
