+++ src/include/Vlucht.hpp
@@ -4,10 +4,10 @@
 
 class Vlucht : public Edge{  
     private:
-        static int multiplyCost(int* cost);
+        static float multiplyCost(float* cost);
 
     public:
-        Vlucht(Node* from, Node* to, int cost);
+        Vlucht(Node* from, Node* to, float cost);
         std::string getType() override;
         bool operator==(Edge* rhs) const override;
 };
