+++ src/include/graph.hpp
@@ -15,7 +15,7 @@ class Graph{
             std::vector<Node*> getNodes() const;
             std::vector<Edge*> getEdges() const;
             Edge* getEdgeBetweenNodes(Node* from, Node* to) const;
-            int getCostOfPath(std::vector<Node*> nodesToVisit) const;
+            float getCostOfPath(std::vector<Node*> nodesToVisit) const;
             std::vector<Node*> findShortestPathWithDijkstra(Node* start, Node* end);
 
             friend bool operator==(std::vector<Node*> lhs, std::vector<Node*>rhs);
